StatsBadge component rewritten to TS
