+++ src/components/StatsBadge/StatsBadge.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import StatsBadge from '@components/StatsBadge'
+import { StatsBadge } from '@components/StatsBadge'
 
 # StatsBadge
