+++ src/components/StatsBadge/StatsBadge.test.tsx
@@ -1,8 +1,7 @@
 import '@testing-library/jest-dom'
 
-import StatsBadge from '@components/StatsBadge'
+import { StatsBadge } from '@components/StatsBadge'
 import { render, screen } from '@testing-library/react'
-import React from 'react'
 
 describe('<StatsBadge /> mount', () => {
   it('is mounted', () => {
