+++ src/components/StatsBadge/index.ts
@@ -1,0 +1,1 @@
+export * from './StatsBadge'
